Lock & Latch

+++ src/com/hasib/java/advance/thread/ThreadLocalExample.java
@@ -3,6 +3,12 @@ package com.hasib.java.advance.thread;
 import java.util.Random;
 import java.util.concurrent.TimeUnit;
 
+/*
+ * Created by S M Al Hasib on 7/23/20, 11:43 PM
+ * Copyright (c) 2020 . All rights reserved.
+ * Last modified 7/23/20, 11:42 PM
+ */
+
 public class ThreadLocalExample {
     public static void main(String[] args) {
         SharedObject sharedObject = new SharedObject();
@@ -32,7 +38,7 @@ public class ThreadLocalExample {
     }
 
     private static void setRandomNumber(SharedObject sharedObject) {
-        sharedObject.setNumber((int) (new Random()).nextInt()*100);
+        sharedObject.setNumber((new Random()).nextInt() * 100);
 
         try {
             TimeUnit.SECONDS.sleep(1);
